Move Amount and Description out of transaction package
