+++ src/main/java/seedu/address/model/Description.java
@@ -1,4 +1,4 @@
-package seedu.address.model.transaction;
+package seedu.address.model;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
